Making SaveManager Complete!
